# src/docs/swagger.yaml
basePath: /v1
definitions:
...
  snippets.CreateSnippetRequest:
    properties:
      content:
        type: string
      language:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
      visibility:
        $ref: '#/definitions/snippets.Visibility'
    type: object
  snippets.Snippet:
    properties:
      content:
        type: string
      created_at:
        type: string
      creator_id:
        type: string
      id:
        type: string
      language:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
      visibility:
        $ref: '#/definitions/snippets.Visibility'
    type: object
  snippets.Visibility:
    enum:
    - public
    - private
    type: string
    x-enum-varnames:
    - VisibilityPublic
    - VisibilityPrivate
...
info:
  contact: {}
  description: sniply_api HTTP API.
  title: sniply_api API
  version: "1.0"
paths:
...
  /snippets:
    get:
      parameters:
      - description: search
        in: query
        name: q
        type: string
      - description: creator id
        in: query
        name: creator
        type: string
      - description: language
        in: query
        name: language
        type: string
      - description: tag
        in: query
        name: tag
        type: string
      - description: visibility
        in: query
        name: visibility
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/snippets.Snippet'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: List snippets
      tags:
      - snippets
    post:
      consumes:
      - application/json
      parameters:
      - description: snippet
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/snippets.CreateSnippetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/snippets.Snippet'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: Create snippet
      tags:
      - snippets
  /snippets/{id}:
    delete:
      parameters:
      - description: snippet id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: Delete snippet
      tags:
      - snippets
    get:
      parameters:
      - description: snippet id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/snippets.Snippet'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: Get snippet by id
      tags:
      - snippets
    put:
      consumes:
      - application/json
      parameters:
      - description: snippet id
        in: path
        name: id
        required: true
        type: string
      - description: snippet
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/snippets.CreateSnippetRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - SessionAuth: []
      - ApiKeyAuth: []
      summary: Update snippet
      tags:
      - snippets
...
securityDefinitions:
  ApiKeyAuth:
    description: 'API key (X-API-Key or Authorization: Bearer)'
    in: header
    name: X-API-Key
    type: apiKey
  SessionAuth:
    description: HttpOnly session cookie
    in: cookie
    name: sniply_session
    type: apiKey
swagger: "2.0"